Created a wrapper class for the interactableBattery and Charging battery classes
