+++ Assets/Prefabs/Hangar/BatterySystem/Battery.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 8424855993406580408}
   - component: {fileID: 1691787280572270736}
   - component: {fileID: 2466530580323449094}
+  - component: {fileID: 4384775771453189675}
   m_Layer: 0
   m_Name: Battery
   m_TagString: Untagged
@@ -93,6 +94,20 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   meshRenderer: {fileID: 4566604538120125738}
+--- !u!114 &4384775771453189675
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 923938306559941887}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 332b6ae005601a44381328caf2450d6a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  BatteryInteractable: {fileID: 8223680205256685485}
+  BatteryCharging: {fileID: 2466530580323449094}
 --- !u!65 &276256587078668969
 BoxCollider:
   m_ObjectHideFlags: 0
