+++ Assets/Scripts/Environment/Hangar/Battery.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 737e1cc8c0998be41ae50cffb1c5c41b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
