+++ Assets/Scripts/Environment/Hangar/Battery/BatteryCharging.cs.meta
