+++ Assets/Scripts/Environment/Hangar/Battery/BatterySlot.cs.meta
