+++ Assets/Scripts/Environment/Hangar/Battery/BatterySpawnPositionManager.cs.meta
