+++ Assets/Scripts/Singletons/DataModelSingletons/HangarManager.cs
@@ -9,8 +9,7 @@ public class HangarManager : MonoBehaviour
     [SerializeField] private GameObject shipInstancePrefab;
 
     public static HangarSlot[] HangarSlots { get; private set; }
-    public static BatteryInteractable[] BatteryInteractables { get; private set; }
-    public static BatteryCharging[] BatteryChargingScripts { get; private set; }
+    public static BatteryWrapper[] BatteryWrappers { get; private set; }
     public static GameObject BatteriesContainer { get; private set; }
 
     public static BatterySpawnPositionManager BatterySpawnPositionManager;
@@ -162,10 +161,10 @@ public class HangarManager : MonoBehaviour
             Debug.LogError("Hangar slots not found");
         }
 
-        BatteryChargingScripts = FindObjectsOfType<BatteryCharging>();
-        if (BatteryChargingScripts.IsNullOrEmpty())
+        BatteryWrappers = FindObjectsOfType<BatteryWrapper>();
+        if (BatteryWrappers.IsNullOrEmpty())
         {
-            Debug.LogError("Battery charging scripts not found");
+            Debug.LogError("Battery wrappers are not found");
         }
 
         BatteriesContainer = GameObject.FindGameObjectWithTag(
@@ -180,51 +179,47 @@ public class HangarManager : MonoBehaviour
         {
             Debug.LogError("Battery spawn position manager not found");
         }
-
-        BatteryInteractables = FindObjectsOfType<BatteryInteractable>();
-        if (BatteryInteractables == null)
-        {
-            Debug.LogError("Battery interactables not found");
-        }
     }
 
     #region Persistence
     public void SaveBatteryData()
     {
-        List<BatteryChargingSaveData> batterySaveData = new List<BatteryChargingSaveData>();
+        List<BatterySaveData> batterySaveData = new List<BatterySaveData>();
 
-        foreach (BatteryCharging batteryCharching in BatteryChargingScripts)
+        foreach (BatteryWrapper batteryWrapper in BatteryWrappers)
         {
-            if (batteryCharching == null)
+            if (batteryWrapper == null)
             {
                 continue;
             }
 
-            batterySaveData.Add(new BatteryChargingSaveData()
+            batterySaveData.Add(new BatterySaveData()
             {
-                IsCharged = batteryCharching.IsCharged,
-                //PositionInHangar = battery.transform.position
+                IsCharged = batteryWrapper.BatteryCharging.IsCharged,
+                PositionInHangar = batteryWrapper.BatteryInteractable.transform.position
             });
         }
         string json = JsonHelper.ListToJson(batterySaveData);
-        string folderPath = DataUtils.GetSaveFolderPath(BatteryCharging.FOLDER_NAME);
-        DataUtils.SaveFileAsync(BatteryCharging.FILE_NAME, folderPath, json);
+        string folderPath = DataUtils.GetSaveFolderPath(BatteryWrapper.FOLDER_NAME);
+        DataUtils.SaveFileAsync(BatteryWrapper.FILE_NAME, folderPath, json);
     }
 
     public async static void LoadBatteryDataAsync()
     {
-        string json = await DataUtils.ReadFileAsync(BatteryCharging.FILE_PATH);
-        BatteryChargingSaveData[] batterySaveData = JsonHelper.ArrayFromJson<BatteryChargingSaveData>(json);
+        string json = await DataUtils.ReadFileAsync(BatteryWrapper.FILE_PATH);
+        BatterySaveData[] batterySaveData = JsonHelper.ArrayFromJson<BatterySaveData>(json);
 
         if (!batterySaveData.IsNullOrEmpty())
         {
             for (int i = 0; i < batterySaveData.Length; i++)
             {
-                if (i > BatteryChargingScripts.Length - 1)
+                if (i > BatteryWrappers.Length - 1)
                 {
                     break;
                 }
-                BatteryChargingScripts[i].LoadData(batterySaveData[i]);
+
+                //BatteryWrappers[i].BatteryInteractable.LoadData(batterySaveData[i]);
+                BatteryWrappers[i].BatteryCharging.LoadData(batterySaveData[i]);
             }
         }
     }
